# apps/backend/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Arena API
  description: |
    Backend foundation API for the arena_new multi-tenant ticketing platform.
    This document covers the foundation milestone endpoints: service metadata,
    health/readiness probes, Prometheus metrics, the transactional echo command
    (which exercises auth, idempotency, audit, and outbox in one transaction),
    and the development-only JWT mint endpoint.
  version: "0.1.0"
  contact:
    name: Arena Platform Team
    email: andreev@abhteam.com
servers:
  - url: http://localhost:8080
    description: Local development server (docker compose)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via POST /v1/dev/token (development mode only,
        requires ENABLE_DEV_AUTH=true) or from the real identity provider
        in production. Pass as "Authorization: Bearer <token>".
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          description: Error details object containing machine-readable code, human-readable message, and correlation IDs.
          required: [code, message, request_id, trace_id]
          properties:
            code:
              type: string
              pattern: '^[a-z][a-z0-9_]+\.[a-z0-9_.]+$'
              description: |
                Machine-readable error code in dotted-namespace format
                (e.g. "http.not_found", "auth.token_expired", "echo.invalid_body").
                Always lowercase; namespace component followed by dot-separated sub-codes.
              example: "auth.token_expired"
            message:
              type: string
              description: Human-readable error description (may be localised)
              example: "authentication required"
            request_id:
              type: string
              description: Per-request correlation ID (also in X-Request-Id response header)
              example: "01J3K8MNTB4KXYZQR5V6W7SHJ2"
            trace_id:
              type: string
              description: Distributed trace ID (also in X-Trace-Id response header)
              example: "4bf92f3577b34da6a3ce929d0e0e4736"
            details:
              type: object
              additionalProperties: true
              description: |
                Optional structured details about the error (analogous to a JSONB object).
                May contain field-level validation errors, conflict context, or other
                machine-readable diagnostic data. Omitted when there are no additional
                details.
              example:
                field: "message"
                reason: "must not be empty"
...
    PaymentProviderConfigItem:
      type: object
      required:
        - id
        - org_id
        - provider
        - mode
        - public_config
        - secret_fields_set
        - status
        - missing_required_fields
        - is_active
        - created_at
        - updated_at
      description: |
        A single payment provider config row for the owning organization.

        Secret values are NEVER returned. `secret_fields_set` lists the
        secret-jsonb keys currently populated (no values). The status field is
        derived from the stored secrets vs. the provider's required-secret
        catalogue (see internal/platform/httpserver/payment_configs.go).
      properties:
        id:
          type: string
          format: uuid
          description: UUIDv7 primary key of the config row
          example: "01929d0e-0e47-7000-8000-000000000600"
        org_id:
          type: string
          format: uuid
          description: Owning organization. Immutable after creation.
          example: "01929d0e-0e47-7000-8000-000000000001"
        provider:
          type: string
          description: |
            Provider slug. Must be one of: stripe, allpay, cloudpayments,
            yookassa, manual.
          example: "stripe"
        mode:
          type: string
          enum: [test, live]
          description: Operating mode for the credential set.
          example: "live"
        provider_account_id:
          type: string
          nullable: true
          description: |
            Optional public account identifier (e.g. Stripe acct_..., AllPay
            merchant id). Safe to return; never a secret.
          example: "acct_1Q2W3E4R5T6Y"
        public_config:
          type: object
          additionalProperties: true
          description: |
            Free-form non-secret provider knobs (e.g. statement_descriptor,
            terminal_id, redirect_url). Always present (default `{}`).
          example:
            statement_descriptor: "ARENA TICKETS"
        secret_fields_set:
          type: array
          items:
            type: string
          description: |
            Names of the secret-jsonb keys currently stored (no values).
            Empty array when no secrets are configured yet.
          example: ["api_key", "webhook_secret"]
        status:
          type: string
          enum: [configured, missing_required_fields]
          description: |
            `configured` when every required secret for the provider is
            populated; otherwise `missing_required_fields`.
          example: "configured"
        missing_required_fields:
          type: array
          items:
            type: string
          description: |
            Names of the required-secret keys still expected for this
            provider. Empty when status = configured.
          example: []
        is_active:
          type: boolean
          description: |
            Whether the config is enabled for use by checkout. Soft-disabled
            configs remain in the database but are skipped by the payment
            router.
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
    UpdatePaymentProviderConfigRequest:
      type: object
      additionalProperties: false
      description: |
        Request body for PATCH /v1/organizations/{org_id}/payment-configs/{id}.
        Omitted fields leave the existing value unchanged. The `secrets`
        field is a PATCH map: non-empty values overwrite, empty-string
        values delete the corresponding key.
      properties:
        provider_account_id:
          type: string
          nullable: true
          description: Replacement account identifier (empty string clears it).
        public_config:
          type: object
          additionalProperties: true
          description: Replacement public_config jsonb (entire object replaces existing).
        secrets:
          type: object
          additionalProperties:
            type: string
          description: |
            Patch map of secret keys. Non-empty value sets the secret,
            empty string deletes the secret, omitted keys are untouched.
        is_active:
          type: boolean
          description: Replacement is_active flag.

    # Geo schemas — feature #123
...
paths:
...
  /v1/organizations/{org_id}/payment-configs/{id}:
    get:
      operationId: getPaymentProviderConfig
      summary: Get one payment provider config by id
      description: |
        Returns the matching active config row. 404 when the row does not
        exist, was soft-deleted, or belongs to a different organization.
        Requires `payment_config.read`.
      tags: [v1]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Config row
          content:
            application/json:
              schema:
                type: object
                required: [payment_config]
                properties:
                  payment_config:
                    $ref: "#/components/schemas/PaymentProviderConfigItem"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Caller lacks payment_config.read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found or owned by another org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Database not wired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updatePaymentProviderConfig
      summary: Partially update a payment provider config
      description: |
        Patches the config in place. The `secrets` map is a per-key patch:
        non-empty values overwrite, empty-string values delete the secret,
        omitted keys are left untouched. The `status` column is recomputed
        from the post-patch secrets jsonb. Requires `payment_config.write`.
      tags: [v1]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentProviderConfigRequest"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                type: object
                required: [payment_config]
                properties:
                  payment_config:
                    $ref: "#/components/schemas/PaymentProviderConfigItem"
        "400":
          description: Invalid body or invalid public_config jsonb
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Caller lacks payment_config.write
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found or owned by another org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Database not wired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deletePaymentProviderConfig
      summary: Soft-delete a payment provider config
      description: |
        Marks the config deleted (sets deleted_at = now()) and writes an
        audit event inside the same transaction. Requires
        `payment_config.write`.
      tags: [v1]
      security:
        - bearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Config soft-deleted
          content:
            application/json:
              schema:
                type: object
                required: [payment_config, deleted]
                properties:
                  payment_config:
                    $ref: "#/components/schemas/PaymentProviderConfigItem"
                  deleted:
                    type: boolean
                    example: true
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Caller lacks payment_config.write
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found or owned by another org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Database not wired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─────────────────────────────────────────────────────────────────────────
  # Operator-network admin endpoints (features #208–#210)
  # ─────────────────────────────────────────────────────────────────────────
...